Store parsed json to a field with name of @source.component
Different vcap components have different JSON messages, but could contain JSON fields with the same name and different type. This causes Logstash warning and lose of data. So we store parsed JSON to a field named like @source.component so that different components do not intersect in naming of JSON fields.