UAA Redirects to Kibana on wrong domain (#270)
The cf-kibana job's manifest.yml uses the CF apps domain for the Kibana app (i.e. logs.apps.cf.com) rather than the system domain.   But this UAA redirect assumes the Kibana app is available from the system domain.    This means UAA redirects will fail.

It's possible that the alternative behavior would be to push the Kibana CF app to a route on the system domain itself rather than the apps domain (which should be a private domain owned by the CF system org).   That's how we worked around this in our install.